+++ EnhanceQoL/NewSettingsTable.lua
@@ -7,4 +7,10 @@ addon.variables.NewVersionTableEQOL = {
 	["EQOL_UFProfiles"] = true,
 	["EQOL_ufProfileActive"] = true,
 	["EQOL_ufProfileGlobal"] = true,
+
+	-- Loot Toast -> Major Factions Renown Toast Anchor
+	["EQOL_GENERAL"] = true,
+	["EQOL_Loot"] = true,
+	["EQOL_enableMajorFactionsRenownToastAnchor"] = true,
+	["EQOL_majorFactionsRenownToastAnchor"] = true,
 }